Fixes #28: Unify all fixtures.ts files and reorganize files
