--- apps/simple/tests/fixtures.ts
@@ -1,32 +1,0 @@
-import { test as base, chromium, type BrowserContext } from "@playwright/test";
-import path from "path";
-
-const pathToExtension = path.resolve("apps/simple/code");
-
-export const test = base.extend<{
-  context: BrowserContext;
-  extensionId: string;
-}>({
-  // eslint-disable-next-line
-  context: async ({}, use) => {
-    const context = await chromium.launchPersistentContext("", {
-      headless: false,
-      args: [
-        `--disable-extensions-except=${pathToExtension}`,
-        `--load-extension=${pathToExtension}`,
-      ],
-    });
-    await use(context);
-    await context.close();
-  },
-  extensionId: async ({ context }, use) => {
-    // for manifest v3:
-    let [serviceWorker] = context.serviceWorkers();
-    if (!serviceWorker)
-      serviceWorker = await context.waitForEvent('serviceworker');
-
-    const extensionId = serviceWorker.url().split('/')[2];
-    await use(extensionId);
-  },
-});
-export const expect = test.expect;
